Create PefectSquares
PerfectSquares
Starting off with a given positive integer, in this case n=12, this is where I print it out in the code. The base comes with a linkedlist, that is with the sumList (this is where it all begins) which finds the sum to n. from here I created a temp so that’s where n would go, which is also the recursion part for the perfect squares, getting the sum from the list. I created an iteration method called Sum. Its own Boolean method is PerfectSquare, to see if n is a perfect, and finally the explanation with its own method which I had to do a lot of debugging at times! But I got it done.